// src/types/messageDTO.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Hopital Management System Api
 * OpenAPI spec version: v1
 */
import type { TicketDTO } from './ticketDTO';
import type { UserProfileDTO } from './userProfileDTO';

export interface MessageDTO {
    /** @nullable */
    content?: string | null;
    createdAt?: string;
    isRead?: boolean;
    messageId?: string;
    ticket?: TicketDTO;
    ticketId?: string;
    user?: UserProfileDTO;
    /** @nullable */
    userId?: string | null;
}